Move gearbox submission to a Celery task
